tests/wpt/web-platform-tests/bluetooth/service/getCharacteristics/gen-characteristic-not-found-with-uuid.https.html

<!-- Generated by //third_party/WebKit/LayoutTests/bluetooth/generate.py -->
<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/bluetooth/resources/bluetooth-helpers.js"></script>
<script>
'use strict';
const test_desc = 'Request for absent characteristics with UUID. ' +
    'Reject with NotFoundError.';

bluetooth_test(() => getEmptyHealthThermometerService()
    .then(({service}) => assert_promise_rejects_with_message(
        service.getCharacteristics('battery_level'),
        new DOMException(
            `No Characteristics matching UUID ${battery_level.uuid} found ` +
            `in Service with UUID ${health_thermometer.uuid}.`,
            'NotFoundError'))),
    test_desc);

</script>
